+++ packages/runner/customer-os-data-upkeeper/service/invoice_service.go
@@ -385,7 +385,6 @@ func (s *invoiceService) updateInvoiceStatusToOverdue() {
 			return // stop processing
 		}
 	}
-
 }
 
 func (s *invoiceService) updateInvoiceStatusToOnHold() {
