+++ packages/server/customer-os-common-module/services/agent_capability/capability_invoice_send_voided_notification.go
@@ -13,68 +13,68 @@ import (
 	"github.com/customeros/customeros/packages/server/customer-os-common-module/tracing"
 )
 
-type SendInvoiceViaEmailCapability struct {
+type SendInvoiceVoidedNotificationCapability struct {
 	invoiceService interfaces.InvoiceService
 }
 
-func NewSendInvoiceViaEmailCapability(invoiceService interfaces.InvoiceService) *SendInvoiceViaEmailCapability {
-	return &SendInvoiceViaEmailCapability{
+func NewSendInvoiceVoidedNotificationCapability(invoiceService interfaces.InvoiceService) *SendInvoiceVoidedNotificationCapability {
+	return &SendInvoiceVoidedNotificationCapability{
 		invoiceService: invoiceService,
 	}
 }
 
 // Compile-time interface checks
 var (
-	_ interfaces.AgentCapability[SendInvoiceViaEmailInput, SendInvoiceViaEmailOutput, postgres_entity.NoConfig] = (*SendInvoiceViaEmailCapability)(nil)
+	_ interfaces.AgentCapability[SendInvoiceVoidedNotificationInput, SendInvoiceVoidedNotificationOutput, postgres_entity.NoConfig] = (*SendInvoiceVoidedNotificationCapability)(nil)
 )
 
-func (c *SendInvoiceViaEmailCapability) Type() enum.AgentCapability {
-	return enum.CapabilitySendInvoiceViaEmail
+func (c *SendInvoiceVoidedNotificationCapability) Type() enum.AgentCapability {
+	return enum.CapabilitySendInvoiceVoidedNotification
 }
 
-func (c *SendInvoiceViaEmailCapability) Name() string {
-	return "Send invoice via email"
+func (c *SendInvoiceVoidedNotificationCapability) Name() string {
+	return "Send invoice voided notification"
 }
 
-func (c *SendInvoiceViaEmailCapability) NewInput() SendInvoiceViaEmailInput {
-	return SendInvoiceViaEmailInput{}
+func (c *SendInvoiceVoidedNotificationCapability) NewInput() SendInvoiceVoidedNotificationInput {
+	return SendInvoiceVoidedNotificationInput{}
 }
 
-func (c *SendInvoiceViaEmailCapability) NewConfig() postgres_entity.NoConfig {
+func (c *SendInvoiceVoidedNotificationCapability) NewConfig() postgres_entity.NoConfig {
 	return postgres_entity.NoConfig{}
 }
 
-func (c *SendInvoiceViaEmailCapability) DefaultConfig() any {
+func (c *SendInvoiceVoidedNotificationCapability) DefaultConfig() any {
 	config := c.NewConfig()
 	return &config
 }
 
-func (c *SendInvoiceViaEmailCapability) ValidateInput(input SendInvoiceViaEmailInput) error {
+func (c *SendInvoiceVoidedNotificationCapability) ValidateInput(input SendInvoiceVoidedNotificationInput) error {
 	if input.InvoiceID == "" {
 		return errors.New("InvoiceID required")
 	}
 	return nil
 }
 
-func (c *SendInvoiceViaEmailCapability) ValidateConfig(postgres_entity.NoConfig) error {
+func (c *SendInvoiceVoidedNotificationCapability) ValidateConfig(postgres_entity.NoConfig) error {
 	return nil
 }
 
-type SendInvoiceViaEmailInput struct {
+type SendInvoiceVoidedNotificationInput struct {
 	InvoiceID string `json:"invoiceId"`
 }
 
-type SendInvoiceViaEmailOutput struct{}
+type SendInvoiceVoidedNotificationOutput struct{}
 
-func (c *SendInvoiceViaEmailCapability) Execute(ctx context.Context, executionContainer interfaces.TypedExecutionContainer[SendInvoiceViaEmailInput, postgres_entity.NoConfig]) (enum.CapabilityExecutionStatus, SendInvoiceViaEmailOutput, error) {
-	span, ctx := opentracing.StartSpanFromContext(ctx, "SendInvoiceViaEmailCapability.Execute")
+func (c *SendInvoiceVoidedNotificationCapability) Execute(ctx context.Context, executionContainer interfaces.TypedExecutionContainer[SendInvoiceVoidedNotificationInput, postgres_entity.NoConfig]) (enum.CapabilityExecutionStatus, SendInvoiceVoidedNotificationOutput, error) {
+	span, ctx := opentracing.StartSpanFromContext(ctx, "SendInvoiceVoidedNotificationCapability.Execute")
 	defer span.Finish()
 	tracing.TagComponentService(span)
 	tracing.TagTenant(span, common.GetTenantFromContext(ctx))
 	tracing.LogObjectAsJson(span, "input", executionContainer.InputData)
 	tracing.LogObjectAsJson(span, "config", executionContainer.ConfigData)
 
-	result := SendInvoiceViaEmailOutput{}
+	result := SendInvoiceVoidedNotificationOutput{}
 
 	if err := c.ValidateInput(executionContainer.InputData); err != nil {
 		tracing.TraceErr(span, errors.Wrap(err, "invalid input"))
@@ -99,16 +99,7 @@ func (c *SendInvoiceViaEmailCapability) Execute(ctx context.Context, executionCo
 		return enum.CapabilityExecutionCompleted, result, nil
 	}
 
-	// TODO alexb check if current agent has payment enabled. if not do not generate payment link and send extra param
-	// TODO cont.. in send pay notification to choose no pay link email template
-
-	err = c.invoiceService.GenerateNewPaymentLink(ctx, executionContainer.InputData.InvoiceID)
-	if err != nil {
-		tracing.TraceErr(span, err)
-		// Do not stop capability execution if payment link generation fails
-	}
-
-	err = c.invoiceService.SendPayInvoiceNotification(ctx, executionContainer.InputData.InvoiceID)
+	err = c.invoiceService.SendVoidedInvoiceNotification(ctx, invoice.Id)
 	if err != nil {
 		tracing.TraceErr(span, err)
 		return enum.CapabilityExecutionCompleted, result, err // failed send invoice email is not a blocker, since a new attempt will be made automatically by cron
